Refactor for a bit of styling
