+++ src/components/Results.scss
@@ -27,7 +27,3 @@
     display: flex;
   }
 }
-
-.screen-title {
-  margin-bottom: 40px;
-}
