+++ src/components/common/RowLink.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+function RowLink(props) {
+  return (
+    <Link {...{ ...props, className: `${props.className} row` }}>
+      {props.children}
+    </Link>
+  )
+};
+
+export default RowLink;
