+++ src/components/common/ScreenTitle.scss
@@ -1,0 +1,3 @@
+.screen-title {
+  margin-bottom: 40px;
+}
